implement part 2 of the requirements

+++ db/migrate/20200214024501_add_composed_of_to_room_units.rb
@@ -1,5 +1,5 @@
 class AddComposedOfToRoomUnits < ActiveRecord::Migration[5.2]
   def change
-    add_column :room_units, :part_of_room_id, :integer
+    add_column :room_units, :part_of_room_id, :bigint
   end
 end
